board listArray to populate

+++ public/src/main.js
@@ -153,11 +153,13 @@ if (window.location.pathname.endsWith("/board")) {
         const confirmDelete = confirm("Are you sure you want to delete?");
         if (confirmDelete) {
             const element = document.querySelector(".isDragging");
-            const listId = element.id;
             element.remove();
-            yield fetch(`${listsAPI}/${listId}`, {
-                method: "DELETE",
-            }).catch((error) => console.error(error));
+            if (element.classList.contains("boardContainer__main__list")) {
+                const listId = element.id;
+                yield fetch(`${listsAPI}/${listId}`, {
+                    method: "DELETE",
+                }).catch((error) => console.error(error));
+            }
             currentBoard.update();
         }
     }));
